# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/send-otp": {
      "post": {
        "operationId": "AuthController_sendOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendOtpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OTP sent to email.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        },
        "summary": "Send OTP for a specific context (e.g. REGISTER, WITHDRAW)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/verify-otp": {
      "post": {
        "operationId": "AuthController_verifyOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OTP verified successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid or expired OTP"
          }
        },
        "summary": "Verify OTP and get a challenge token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully registered.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Email already in use or validation error"
          }
        },
        "summary": "Register a new user",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User successfully logged in.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid credentials"
          }
        },
        "summary": "Login and get tokens",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tokens successfully refreshed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or expired refresh token"
          }
        },
        "summary": "Refresh access token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Logged out successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Logout",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/change-password": {
      "put": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Incorrect old password"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change user password",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/notifications": {
      "get": {
        "operationId": "NotificationsController_getNotifications",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return a list of notifications.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedNotificationResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user notifications history",
        "tags": [
          "Notifications"
        ]
      }
    },
...
  },
  "info": {
    "title": "Task Bounty Platform API",
    "description": "The Nền tảng Quản lý công việc và Giao việc tự do API description.\n\n### Real-time Events (WebSocket)\nThis platform provides real-time updates via Socket.io.\n- **Endpoint**: `ws://<host>:<port>/notifications`\n- **Transports**: `['websocket', 'polling']`\n\n#### Events to Listen For\n1. **`risk-alert`**: Emitted when the AI detects a high risk score for a task.\n   - Payload: `{ pmId: string, message: string }`\n2. **`balance-warning`**: Emitted when the PM's escrow balance drops below a required threshold.\n   - Payload: `{ pmId: string, message: string }`\n",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "SendOtpDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "test@example.com"
          },
          "context": {
            "type": "string",
            "example": "REGISTER",
            "description": "Context of the OTP, e.g. REGISTER, WITHDRAW, DEPOSIT"
          }
        },
        "required": [
          "email",
          "context"
        ]
      },
      "MessageResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Operation completed successfully"
          }
        },
        "required": [
          "success",
          "message"
        ]
      },
      "VerifyOtpDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "test@example.com"
          },
          "code": {
            "type": "string",
            "example": "123456"
          },
          "context": {
            "type": "string",
            "example": "REGISTER"
          }
        },
        "required": [
          "email",
          "code",
          "context"
        ]
      },
      "ChallengeResponseDto": {
        "type": "object",
        "properties": {
          "challengeToken": {
            "type": "string",
            "example": "eyJh...",
            "description": "A short-lived JWT token to prove OTP verification"
          }
        },
        "required": [
          "challengeToken"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "username": {
            "type": "string",
            "example": "johndoe"
          },
          "email": {
            "type": "string",
            "example": "test@example.com"
          },
          "password": {
            "type": "string",
            "example": "password123"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "username",
          "email",
          "password"
        ]
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-1234"
          },
          "email": {
            "type": "string",
            "example": "test@example.com"
          },
          "walletAddress": {
            "type": "string",
            "example": "0x123abc..."
          }
        },
        "required": [
          "id",
          "email",
          "walletAddress"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "example": "eyJh..."
          },
          "refresh_token": {
            "type": "string",
            "example": "eyJh..."
          },
          "user": {
            "$ref": "#/components/schemas/UserDto"
          }
        },
        "required": [
          "access_token",
          "refresh_token",
          "user"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "identifier": {
            "type": "string",
            "example": "test@example.com or johndoe"
          },
          "password": {
            "type": "string",
            "example": "password123"
          }
        },
        "required": [
          "identifier",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "eyJh..."
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "oldPassword": {
            "type": "string",
            "example": "old_password123"
          },
          "newPassword": {
            "type": "string",
            "example": "new_password123"
          }
        },
        "required": [
          "oldPassword",
          "newPassword"
        ]
      },
...
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "lastPage": {
            "type": "number"
          }
        },
        "required": [
          "total",
          "page",
          "limit",
          "lastPage"
        ]
      },
...
      "NotificationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-1234"
          },
          "userId": {
            "type": "string",
            "example": "uuid-user-1234"
          },
          "message": {
            "type": "string",
            "example": "Your task has been approved."
          },
          "isRead": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "example": "2026-07-30T10:00:00Z"
          }
        },
        "required": [
          "id",
          "userId",
          "message",
          "isRead",
          "createdAt"
        ]
      },
      "PaginatedNotificationResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotificationResponseDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
    }
  }
}